tn48m: drop fan script
Currently, the script is using the ADT7343 fan controllers
auto mode.

Issue with that mode is that it is really agressive and does
not offer fine grained control in the end causing high
PWM and RPM values along with the noise.

So, lets drop this in favor of using lm-sensors fancontrol.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>