armada7040-prestera: update kernel to 5.14.6
A long overdue kernel update from 5.12.12 to 5.14.6.
It allows us to get rid of upstreamed patches.

It also finally carries the updated Prestera driver to the
v3.0.1 that is patched to run on 5.14 instead of 5.10.

Also, fixup the fancontrol enablement as now systemd
reads the preset ordering in alphabetical ordering
hence why it now newer gets enabled as the default
99-default.preset which disables all units by default
has higher priority.
So, simply move it to 98-fancontrol.preset.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>