Makefile: use stage3-amd64-openrc and update to 20211018
Gentoo has renamed the default stage3-amd64 builds to stage3-amd64-openrc
as they now provide an systemd variant as well.

So, switch to using the stage3-amd64-openrc tag and update to build
20211018 as the stage3-amd64 has not been updated since June 30th.

It also doesnt contain GCC11 which is now the default so it causes a
rebuild of the host GCC as well which takes a long time and beats the
purpose of having prebuild host tools.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>